pyproject.toml
[project]
name = "tts-llm-podcast-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "edge-tts>=7.2.7",
    "fastapi>=0.129.0",
    "google-generativeai>=0.8.6",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
